Move core metrics classes to core package
