+++ solr/core/src/java/org/apache/solr/metrics/prometheus/core/SolrCoreHighlighterMetric.java
@@ -1,7 +1,8 @@
-package org.apache.solr.metrics.prometheus;
+package org.apache.solr.metrics.prometheus.core;
 
 import com.codahale.metrics.Counter;
 import com.codahale.metrics.Metric;
+import org.apache.solr.metrics.prometheus.SolrPrometheusCoreRegistry;
 
 public class SolrCoreHighlighterMetric extends SolrCoreMetric {
   public static final String CORE_HIGHLIGHER_METRICS = "solr_metrics_core_highlighter_requests";
@@ -22,7 +23,7 @@ public class SolrCoreHighlighterMetric extends SolrCoreMetric {
   }
 
   @Override
-  void toPrometheus(SolrPrometheusCoreRegistry solrPrometheusCoreRegistry) {
+  public void toPrometheus(SolrPrometheusCoreRegistry solrPrometheusCoreRegistry) {
     if (dropwizardMetric instanceof Counter) {
       solrPrometheusCoreRegistry.exportCounter(
           (Counter) dropwizardMetric, CORE_HIGHLIGHER_METRICS, labels);
