+++ solr/core/src/java/org/apache/solr/metrics/prometheus/core/SolrCoreIndexMetric.java
@@ -1,7 +1,8 @@
-package org.apache.solr.metrics.prometheus;
+package org.apache.solr.metrics.prometheus.core;
 
 import com.codahale.metrics.Gauge;
 import com.codahale.metrics.Metric;
+import org.apache.solr.metrics.prometheus.SolrPrometheusCoreRegistry;
 
 public class SolrCoreIndexMetric extends SolrCoreMetric {
   public static final String CORE_INDEX_METRICS = "solr_metrics_core_index_size_bytes";
@@ -17,7 +18,7 @@ public class SolrCoreIndexMetric extends SolrCoreMetric {
   }
 
   @Override
-  void toPrometheus(SolrPrometheusCoreRegistry solrPrometheusCoreRegistry) {
+  public void toPrometheus(SolrPrometheusCoreRegistry solrPrometheusCoreRegistry) {
     if (dropwizardMetric instanceof Gauge) {
       if (metricName.endsWith("sizeInBytes")) {
         solrPrometheusCoreRegistry.exportGauge(
